make rpc/packets/PrecommitBlock only to use DataBlock.meta
