+++ src/query/datablocks/Cargo.toml
@@ -27,6 +27,8 @@ parking_lot = "0.12.1"
 pin-project-lite = "0.2.9"
 primitive-types = "0.12.0"
 regex = "1.6.0"
+serde = { workspace = true }
+typetag = "0.2.3"
 
 [dev-dependencies]
 pretty_assertions = "1.3.0"
